requirements.txt
# Core Flask application
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-wtf==1.2.1
flask-limiter==3.5.0
authlib==1.2.1
email-validator==2.0.0
bleach==6.0.0

# Web scraping and HTTP requests
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.7

# Database
psycopg2-binary==2.9.7
# sqlite3 is built into Python, no need to install

# Data processing and utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Optional: For advanced web scraping (uncomment if needed)
# selenium==4.15.0
# chromedriver-binary==118.0.5993.70.0

# Payment processing
stripe==7.0.0

# Production server
gunicorn==21.2.0

# Environment and configuration
python-decouple==3.8